adds an extra security layer to api endpoints

+++ app/api/tickets/batches/[id]/route.ts
@@ -1,9 +1,18 @@
+import { getUserRole } from "@/app/lib/dal";
 import { prisma } from "@/app/lib/db";
 import { ErrorResponse, GenericResponse } from "@/utils/responses";
 import { NextRequest } from "next/server";
 
 export async function DELETE(req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
     try {
+        if (req.method !== "DELETE") 
+            return ErrorResponse("method not allowed", 405);
+                
+        const role = await getUserRole()??"";
+        
+        if (role !== "PRODUCER")
+            return ErrorResponse("forbidden", 403);
+
         const batchId = (await params).id;
         
         const deleted = await prisma.ticketBatch.delete({where: {
